requirements.txt
flask==0.9
requests==1.1.0
python-dateutil==1.5
pymongo==2.4.2